requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.7
django-appconf==1.0.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-imagekit==4.0.2
django-rest-swagger==2.2.0
django-split-settings==0.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
mysqlclient==1.3.12
openapi-codec==1.3.2
pilkit==2.0
Pillow==5.2.0
pluggy==0.6.0
py==1.5.4
PyJWT==1.6.4
pytest==3.6.3
pytest-django==3.3.3
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
